Simplify usage of smoke test by having it do port forwarding
